Refactor AnnotationsUtils class - Move AnnotationsUtils and SchemaUtils to config directory - Move AnnotationsUtils' Response-dependent methods to ResponseAnnotationsUtils class
